SolutionBug Delete 수정 및 정리

+++ teamKMS/src/main/java/com/devworker/kms/repo/solution/SolutionBugRepo.java
@@ -7,9 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 
-public interface SolutionBugRepo extends PagingAndSortingRepository<SolutionBugDao, Integer> {
-//	@Query ("select b from BoardDao b, SolutionBugDao t where b.board_id = t.board_id")
-//	@Query ("select b from BoardDao b")	
+public interface SolutionBugRepo extends PagingAndSortingRepository<SolutionBugDao, Long> {
 	@Query ("select t from SolutionBugDao t")	
     Page<SolutionBugDao> getBug(@Param("menuId")int menuId, Pageable pageable);
 }
